GPG-629 Removed the unused ActionHubController.New and views
