+++ GenderPayGap.WebUI/Views/ActionHub/ActionHubLayout.cshtml
@@ -1,0 +1,63 @@
+﻿@using GovUkDesignSystem.GovUkDesignSystemComponents
+@using GovUkDesignSystem
+
+@{
+    ViewBag.Title = "Actions to close the gender pay gap - Gender pay gap service";
+    Layout = "~/Views/GovUkFrontend/GovUkFrontendLayout.cshtml";
+}
+
+@section BeforeMain {
+    @{
+        var crumbs = new List<CrumbViewModel>
+        {
+            new CrumbViewModel
+            {
+                Text = "Home",
+                Href = Url.Action("Index", "Viewing")
+            },
+            new CrumbViewModel
+            {
+                Text = "Actions to close the gender pay gap"
+            }
+        };
+    }
+
+    @(Html.GovUkBreadcrumbs(new BreadcrumbsViewModel
+    {
+        Crumbs = crumbs
+    }))
+}
+
+<partial name="Sections/Header" />
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <h4 class="govuk-caption-m">Contents</h4>
+
+        <ul class="govuk-list guidance-list-with-dashes">
+            <li>
+                <a class="govuk-link govuk-body-s" href="@Url.Action("Overview", "ActionHub")">Overview</a>
+            </li>
+            <li>
+                <a class="govuk-link govuk-body-s" href="@Url.Action("Effective", "ActionHub")">Effective actions</a>
+            </li>
+            <li>
+                <a class="govuk-link govuk-body-s" href="@Url.Action("Promising", "ActionHub")">Promising actions</a>
+            </li>
+            <li>
+                <a class="govuk-link govuk-body-s" href="@Url.Action("MixedResult", "ActionHub")">Actions with mixed results</a>
+            </li>
+        </ul>
+    </div>
+</div>
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <div class="govuk-!-margin-top-4">
+            @RenderBody()
+        </div>
+    </div>
+    <div class="govuk-grid-column-one-third">
+        <partial name="Sections/RelatedContent" />
+    </div>
+</div>
