+++ GenderPayGap.WebUI/Views/ActionHub/Sections/Header.cshtml
@@ -1,0 +1,26 @@
+﻿<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <span class="govuk-caption-l">Guidance</span>
+        <h1 class="govuk-heading-xl">Actions to close the gender pay gap</h1>
+    </div>
+</div>
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <p class="govuk-body-l">
+            Actions that are likely to improve recruitment and progression of women and reduce the gender pay gap.
+        </p>
+        <hr class="govuk-section-break govuk-section-break--xl govuk-section-break--visible govuk-!-margin-bottom-2">
+        <div class="govuk-!-margin-bottom-6">
+            <p class="govuk-body-s govuk-!-margin-bottom-0">
+                Published 1 December 2020
+            </p>
+
+            <p class="govuk-body-s govuk-!-margin-bottom-0">
+                From: <a class="govuk-link" href="https://www.gov.uk/government/organisations/government-equalities-office" rel="noopener">
+                    Government Equalities Office
+                </a>
+            </p>
+        </div>
+    </div>
+</div>
