+++ GenderPayGap.WebUI/Views/ActionHub/Sections/RelatedContent.cshtml
@@ -1,0 +1,59 @@
+﻿@using GenderPayGap.Core
+
+<div class="guidance-related-content">
+    <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible related-content-section-break">
+
+    <h3 class="govuk-heading-s">Related content</h3>
+
+    <ul class="govuk-list related-content-list">
+        <li>
+            <a class="govuk-link govuk-body-s" href="/public/assets/pdf/action-plan-guidance.pdf" rel="download">
+                Four steps to developing a gender pay gap action plan
+                <span class="govuk-visually-hidden"> (pdf)</span>
+            </a>
+        </li>
+        <li>
+            <a class="govuk-link govuk-body-s" href="@Url.Action("EightWaysOverview", "EightWaysGuidance")">
+                Eight ways to understand your organisation's gender pay gap
+            </a>
+        </li>
+        <li>
+            <a class="govuk-link govuk-body-s" href="@Url.Action("SearchResults", "Viewing")">
+                Search and compare employer gender pay gaps
+            </a>
+        </li>
+        @if (FeatureFlagHelper.IsFeatureEnabled(FeatureFlag.ReportingStepByStep))
+        {
+            <li>
+                <a class="govuk-link govuk-body-s" href="@Url.Action("StepByStepStandalone", "ReportingStepByStep")">
+                    Prepare and report your gender pay gap: step-by-step guidance for employers
+                </a>
+            </li>
+        }
+    </ul>
+    <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
+    <div>
+        <h4 class="govuk-caption-m">Collection</h4>
+        <a class="govuk-link govuk-body-s" href="@Global.GpgGuidanceLink">
+            Gender pay gap reporting
+        </a>
+    </div>
+</div>
+<hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
+<div class="guidance-related-content">
+
+    <h3 class="govuk-heading-s">Further reading</h3>
+
+    <ul class="govuk-list related-content-list">
+        <li class="govuk-body-s">
+            What Works: Gender equality by design<br />
+            by Iris Bohnet<br />
+            Harvard University Press, 2016
+        </li>
+        <li class="govuk-body-s">
+            Why diversity programs fail<br />
+            by F. Dobbin and A. Kalev<br />
+            Harvard Business Review, 2016
+        </li>
+    </ul>
+</div>
