--- GenderPayGap.WebUI/Views/ActionHubNew/Index.cshtml
@@ -1,79 +1,0 @@
-﻿@{
-    Layout = "~/Views/Dfe/Layouts/template.cshtml";
-    ViewBag.Title = "Actions to close the gap";
-    ViewBag.pageDescription = "";
-    ViewBag.pageClass = "actions-to-close-gap-page";
-}
-
-<div class="grid-row">
-    <div class="column-two-thirds">
-        <h1 class="heading-xlarge">
-            Actions to close the gender pay gap
-        </h1>
-    </div>
-</div>
-
-<div class="grid-row">
-    <div class="column-two-thirds">
-        <div class="page-contents js-page-contents" style="">
-            <h2 class="page-contents__title">Contents</h2>
-            <ul class="page-contents__list">
-                <li>
-                    <a asp-controller="ActionHub" asp-action="OverView">Overview</a>
-                </li>
-                <li>
-                    <a asp-controller="ActionHub" asp-action="Leadership">Actions: Leadership and accountability</a>
-                </li>
-                <li>
-                    <a asp-controller="ActionHub" asp-action="Hiring">Actions: Hiring and selection</a>
-                </li>
-                <li>
-                    <a asp-controller="ActionHub" asp-action="Talent">Actions: Talent management, learning &amp; development</a>
-                </li>
-                <li>
-                    <a asp-controller="ActionHub" asp-action="Workplace">Actions: Workplace flexibility</a>
-                </li>
-                <li>
-                    <a asp-controller="ActionHub" asp-action="Reading">Further reading</a>
-                </li>
-            </ul>
-        </div>
-    </div>
-</div>
-
-<div class="grid-row">
-    <div class="column-two-thirds">
-        @RenderBody()
-    </div>
-    <div class="column-one-third">
-        <div class="gem-c-related-navigation" style="margin-top: 35px">
-            <nav role="navigation" class="gem-c-related-navigation__nav-section font-xsmall" aria-labelledby="related-nav-related_items" data-module="toggle">
-                <h3 class="heading-small" id="related-nav-related_items">Related content</h3>
-                <ul class="gem-c-related-navigation_link-list" data-module="track-click">
-                    <li class="gem-c-related-navigation__link">
-                        <a class="gem-c-related-navigation__section-link--other" data-track-category="Related Link Clicked" href="/viewing/search-results?t=2">Find and compare employer's gender pay gaps</a>
-                    </li>
-                    <li class="gem-c-related-navigation__link">
-                        <a class="gem-c-related-navigation__section-link--other" data-track-category="Related Link Clicked" href="/public/assets/pdf/understand-your-gender-pay-gap.pdf">Eight ways to understand your organisation’s gender pay gap</a>
-                    </li>
-                    <li class="gem-c-related-navigation__link">
-                        <a class="gem-c-related-navigation__section-link--other" data-track-category="Related Link Clicked" href="/public/assets/pdf/action-plan-guidance.pdf">Four steps to developing a gender pay gap action plan</a>
-                    </li>
-                    <li class="gem-c-related-navigation__link">
-                        <a class="gem-c-related-navigation__section-link--other" data-track-category="Related Link Clicked" href="/public/assets/pdf/gender-pay-gap-explained.pdf">Gender pay gap explained</a>
-                    </li>
-                </ul>
-            </nav>
-        </div>
-    </div>
-</div>
-
-@section Scripts {
-    <script>
-        (function () {
-            "use strict";
-            GOVUK.details.addDetailsPolyfill();
-            new GOVUK.Modules.AccordionWithDescriptions().start($('.js-accordion-with-descriptions'));
-        }());
-    </script>
-}
