--- GenderPayGap.WebUI/Views/ActionHubNew/Leadership.cshtml
@@ -1,132 +1,0 @@
-﻿@{
-    ViewBag.Title = "Leadership and accountability actions to close the gender pay gap - GOV.UK";   
-    ViewBag.pageDescription = "Find the most effective leadership and accountability actions to close the gender pay gap";
-    Layout = "Index.cshtml";
-    var index = 1;
-}
-
-<h2 class="heading-large">
-    Leadership and accountability
-</h2>
-<p>
-    Leaders play a key role in driving gender equality in their organisations. The first concrete thing leaders can do is to set specific, ambitious targets. The second thing is to create strong accountability mechanisms to ensure that every hiring manager and people manager feels accountable for diversity outcomes in their team.
-</p>
-<div class="action-step">
-    <i class="icon icon-effective"></i>
-    <h3 class="heading-small">
-        Effective actions
-    </h3>
-    <ul class="list list-bullet">
-        <li>Set internal targets for gender representation and equality</li>
-        <li>
-            Appoint diversity managers and/or diversity task forces
-        </li>
-    </ul>
-    <expand id="MoreEffectiveActions" summary="More on these effective actions">
-        <h3 class="heading-small">
-            Set internal targets for gender representation and equality
-        </h3>
-        <p>
-            The first thing many organisations do is set goals. Unfortunately general goals such as "we will reduce our organisation’s gender pay gap" are not effective unless they are accompanied by specific goals. High-level goals do not spur action or assign responsibility. Targets are most successful when they are:
-        </p>
-        <ul class="list list-bullet">
-            <li>Specific and clear. What are you aiming to change and how big will the change be?</li>
-            <li>Time-bound. By when will the goal be achieved?</li>
-            <li>Challenging but realistic</li>
-            <li>
-                Public. Organisations should consider including their targets in the action plans they publish on the Gender Pay Gap reporting website, and in other public locations. Making a public commitment makes it more likely that the target will be achieved
-            </li>
-            <li>Monitored. Progress towards the goals should be tracked and reviewed regularly</li>
-        </ul>
-        <h3 class="heading-small">
-            Appoint diversity managers and/or diversity task forces
-        </h3>
-        <p>
-            Diversity managers and diversity task forces hold all parts of the organisation accountable for diversity and inclusion efforts and outcomes. Having a diversity manager is associated with better representation of women in organisations.
-        </p>
-        <p>
-            However, simply appointing a diversity manager or task force is not enough. Diversity managers and task forces should be able to review hiring and progression decisions and ask for justifications for those decisions. This creates accountability. When people know their decisions may be reviewed by a senior diversity manager or task force, they pay closer attention to the information they are basing their decisions on, and make less biased decisions.
-        </p>
-        <p>
-            Diversity managers and task forces should:
-        </p>
-        <ul class="list list-bullet">
-            <li>
-                Have a senior/executive role within the organisation. They need to have enough power and influence to hold people accountable for making progress on equality and diversity goals
-            </li>
-            <li>
-                Have visibility of internal data so they can track progress and outcomes
-            </li>
-            <li>
-                Be in a position to ask for more information on why decisions were made
-            </li>
-            <li>
-                Be empowered to develop and implement diversity strategies and policies
-            </li>
-        </ul>
-    </expand>
-</div>
-
-<Accordion>
-    <Section id="evidence-categories" title="Evidence categories">
-        <div class="action-step">
-            <i class="icon icon-effective"></i>
-            <h3 class="heading-small">
-                Effective actions
-            </h3>
-            <p>
-                There is strong evidence that shows these actions are effective, and that they are worth implementing. We recommend prioritising these actions above the actions in the categories below.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-promising"></i>
-            <h3 class="heading-small">
-                Promising actions
-            </h3>
-            <p>
-                These actions are promising but they still need further research to improve the evidence of their effectiveness and how best to implement them. The government has policies in place supporting some of these actions to improve women's participation in the workforce. The government is evaluating their effectiveness and we recommend that employers evaluate their actions too.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-mixed"></i>
-            <h3 class="heading-small">
-                Actions with mixed results
-            </h3>
-            <p>
-                These actions have been shown sometimes to have a positive impact and other times a negative impact. This might be due to how they are implemented or other factors that we do not fully understand yet. As a result, investing in these actions is a more risky use of resources. We cannot, therefore, make a general recommendation that these are good ways to reduce gender inequality.
-            </p>
-        </div>
-    </Section>
-    <Section id="References" title="References">
-        <h3 class="heading-small">
-            Effective action: Set internal targets for gender representation and equality
-        </h3>
-        <p>
-            Mento, A.J., Steel, R.P. & Karren, R.J. (1987). A meta-analytic study of the effects of goal setting on task performance: 1966–1984. Organizational Behaviour and Human Decision Processes, 39(1), 52-83.
-        </p>
-        <p>
-            Epton, T., Currie, S. & Armitage, C.J. (2017). Unique effects of setting goals on behavior change: Systematic review and meta-analysis. Journal of Consulting and Clinical Psychology, 85(12), 1182-1198.
-        </p>
-        <h3 class="heading-small">
-            Effective action: Appoint diversity managers and/or diversity task forces
-        </h3>
-        <p>
-            Dobbin, F., & Kalev, A. (2016). <a class="nowrap" href="https://hbr.org/2016/07/why-diversity-programs-fail" rel="noopener">Why diversity programs fail.</a><br/>Harvard Business Review, 94(7/8), 52-60.
-        </p>
-
-        <p>
-            Castilla, E.J. (2015). Accounting for the gap: A firm study manipulating organizational accountability and transparency in pay decisions. Organization Science, 26(2), 311-333.
-        </p>
-        <p>
-            Dobbin, F., Schrage, D. & Kalev, A. (2015). Rage against the iron cage: The varied effects of bureaucratic personnel reforms on diversity. American Sociological Review, 80(5), 1014-1044.
-        </p>
-    </Section>
-</Accordion>
-
-<Pagination>
-    <Back action="Overview" label="Overview"/>
-    <Next action="Hiring" label="Hiring and selection"/>
-</Pagination>
-<div style="margin-bottom: 25px; font-size: 16px!important">
-    <a rel="download" class="print-link" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf">Print entire guide<span class="visually-hidden"> (pdf)</span></a>
-</div>
