--- GenderPayGap.WebUI/Views/ActionHubNew/Overview.cshtml
@@ -1,108 +1,0 @@
-﻿@{
-    ViewBag.Title = "";   
-    ViewBag.pageDescription = "Find out how employers can close the gender pay gap. See the evidence and download our 'What Works' Guide";
-    Layout = "Index.cshtml";
-}
-
-<h2 class="heading-large">Overview</h2>
-<p>
-    Employers have asked us which actions are likely to improve recruitment and progression of women, and reduce the gender pay gap.
-</p>
-<p>
-    This updated guidance summarises approaches that have been shown to improve outcomes for women. It also indicates those which need more evidence before they can be recommended as widespread approaches. This will help employers create more effective action plans.
-</p>
-<p>
-    This guidance groups actions into four different areas of focus:
-</p>
-<ul class="list list-bullet">
-    <li>Leadership and accountability</li>
-    <li>Hiring and selection</li>
-    <li>Talent management, learning and development</li>
-    <li>Workplace flexibility</li>
-</ul>
-
-<Accordion>
-    <Section id="evidence-categories" title="Evidence categories">
-        <div class="action-step big">
-            <i class="icon icon-effective"></i>
-            <h3 class="heading-small">
-                Effective actions
-            </h3>
-            <p>
-                There is strong evidence that shows these actions are effective, and that they are worth implementing. We recommend prioritising these actions above the actions in the categories below.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-promising"></i>
-            <h3 class="heading-small">
-                Promising actions
-            </h3>
-            <p>
-                These actions are promising but they still need further research to improve the evidence of their effectiveness and how best to implement them. The government has policies in place supporting some of these actions to improve women's participation in the workforce. The government is evaluating their effectiveness and we recommend that employers evaluate their actions too.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-mixed"></i>
-            <h3 class="heading-small">
-                Actions with mixed results
-            </h3>
-            <p>
-                These actions have been shown sometimes to have a positive impact and other times a negative impact. This might be due to how they are implemented or other factors that we do not fully understand yet. As a result, investing in these actions is a more risky use of resources. We cannot, therefore, make a general recommendation that these are good ways to reduce gender inequality.
-            </p>
-        </div>
-    </Section>
-    <Section id="evidence-used" title="What evidence we have used">
-        <p>
-            We focused on gender-related research and evidence, but we also included research on characteristics such as race/ethnicity and sexual orientation to understand what works for improving equality and diversity more generally. This acknowledges that people’s experiences are not determined by a single dimension of identity.
-        </p>
-        <p>
-            For some actions, we have gone beyond gender and other diversity characteristics, drawing on a wider body of behavioural science research to make our recommendations. For example, there are no high quality studies on the impact of setting internal targets on equality in the workplace. We have nonetheless included this as an effective action because goal setting has a consistent effect on behaviour across a range of settings. So where the body of findings is robust and consistent across contexts, we have included the action even without specific evidence from the workplace equality context.
-        </p>
-        <p>
-            This guidance is based on the best evidence currently available. However, highly rigorous evidence is currently scarce in the field of gender equality in the workplace. Our judgement about whether a given action is categorised as "effective", "promising" or "mixed evidence" is based on a holistic assessment of the current evidence. In this document, we cite only a sample of the studies that were used to inform our assessment of the efficacy of a given action as it was not practical to cite them all.
-        </p>
-        <p>
-            This guidance will be updated as the evidence base develops.
-        </p>
-    </Section>
-    <Section id="actions-to-implement" title="Which actions to implement">
-        <p>Every organisation is different and may face different challenges. To know where to focus your efforts, we recommend taking a data-driven approach. </p>
-        <p>Answering the questions in <a rel="track" class="gem-c-related-navigation__section-link--other" data-track-category="Related Link Clicked" href="/public/assets/pdf/understand-your-gender-pay-gap.pdf">Eight ways to understand your organisation’s gender pay gap</a> will help you identify where gender imbalance is occurring and address the specific issues in your organisation.</p>
-        <p>
-            Targeted actions are important but gender equality in an organisation is not achieved by implementing just a few actions - it is driven by a sustained approach that reaches every part of the organisation.
-        </p>
-    </Section>
-</Accordion>
-
-<Pagination>
-    <Next action="Leadership" label="Leadership and accountability"/>
-</Pagination>
-
-<div style="margin-bottom: 25px; font-size: 16px!important">
-    <a rel="download track" class="print-link" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf">Print entire guide<span class="visually-hidden"> (pdf)</span></a>
-</div>
-@*Actions to close the gender pay gap - download PDF*@
-<table class="table-responsive">
-    <tbody>
-    <tr>
-        <td style="width: 100px;">
-            <a download rel="download track" class="gem-c-related-navigation__section-link--other" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf">
-                <img style="border:6px solid #dee0e2" width="78" src="/img/what-works2.png" alt=""/>
-                <span class="visually-hidden">Actions for employers to close the gender pay gap (pdf)</span>
-            </a>
-        </td>
-        <td style="vertical-align: top">
-            <div style="width: 180px; align-content: left;">
-                <a download rel="download track" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf"
-                   style="font-size: 19px !important">
-                    Actions for employers to close the gender pay gap
-                    <span class="visually-hidden"> (pdf)</span>
-                </a>
-            </div>
-            <div class="text-left font-xsmall">
-                <p>PDF, 6.37MB, 24 pages</p>
-            </div>
-        </td>
-    </tr>
-    </tbody>
-</table>
