--- GenderPayGap.WebUI/Views/ActionHubNew/Reading.cshtml
@@ -1,45 +1,0 @@
-﻿@{
-    ViewBag.Title = "Further reading on how to close the gap - GOV.UK";   
-    ViewBag.pageDescription = "Find out more about closing the gender pay gap";
-    Layout = "Index.cshtml";
-}
-
-<h2 class="heading-large">Further reading</h2>
-
-<p>
-    <b>What Works: Gender equality by design</b><br/>
-    by Iris Bohnet<br/>
-    Harvard University Press, 2016
-</p>
-<p>
-    <a data-track-category="Further reading link clicked" href="https://hbr.org/2016/07/why-diversity-programs-fail" rel="noopener track">
-        <b>Why diversity programs fail</b>
-    </a><br/>by F. Dobbin and A. Kalev<br/>Harvard Business Review, 2016
-</p>
-<p>
-    <a data-track-category="Further reading link clicked" href="https://www.cipd.co.uk/knowledge/culture/behaviour/recruitment-report" rel="noopener track">
-        <b>A head for hiring: The behavioural science of recruitment and selection</b>
-    </a><br/>CIPD, 2015
-</p>
-<p>
-    <a data-track-category="Further reading link clicked" href="https://www.gov.uk/government/publications/womens-progression-in-the-workplace-actions-for-employers" rel="noopener track">
-        <b>Women’s Progression in the Workplace: Actions for employers</b>
-    </a><br/>Government Equalities Office, 2019
-</p>
-<p>
-    <a data-track-category="Further reading link clicked" href="https://www.gov.uk/government/publications/family-friendly-policies-actions-for-employers" rel="noopener track">
-        <b>Family friendly policies: Actions for employers</b>
-    </a><br/>Government Equalities Office, 2019
-</p>
-<p>
-    <a data-track-category="Further reading link clicked" href="https://www.womensbusinesscouncil.co.uk/100-ways-to-work/" rel="noopener track">
-        <b>#100WaysToWorkFlexibly</b>
-    </a><br/>Women’s Business Council, 2019
-</p>
-
-<Pagination>
-    <Back action="Workplace" label="Workplace flexibility"/>
-</Pagination>
-<div style="margin-bottom: 25px; font-size: 16px!important">
-    <a rel="download track" class="print-link" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf">Print entire guide<span class="visually-hidden"> (pdf)</span></a>
-</div>
