--- GenderPayGap.WebUI/Views/ActionHubNew/Workplace.cshtml
@@ -1,129 +1,0 @@
-﻿@{
-    ViewBag.Title = "Workplace flexibility actions to close the gender pay gap - GOV.UK";   
-    ViewBag.pageDescription = "Find out how workplace flexibility can help close the gender pay gap";
-    Layout = "Index.cshtml";
-    var index = 1;
-}
-
-<h2 class="heading-large">
-    Workplace Flexibility
-</h2>
-<p>
-    Flexible working arrangements and generous parental leave policies are increasingly important in attracting the best talent. It is important for employers to ‘walk the talk’ and ensure that take up of these policies is genuinely encouraged throughout the organisation.
-</p>
-<div class="action-step">
-    <i class="icon icon-promising"></i>
-    <h3 class="heading-small">
-        Promising actions
-    </h3>
-    <ul class="list list-bullet">
-        <li>
-            Improve workplace flexibility for men and women
-        </li>
-        <li>
-            Encourage the uptake of Shared Parental Leave
-        </li>
-    </ul>
-    <expand id="MorePromisingActions" summary="More on these promising actions">
-        <h3 class="heading-small">
-            Improve workplace flexibility for men and women
-        </h3>
-        <p>
-            It is important to challenge the gender stereotype that it is a women’s role to take on caring responsibilities. Improving workplace flexibility for everyone can enable both women and men to combine work with family and other parts of their lives.
-        </p>
-        <ul class="list list-bullet">
-            <li>
-                Advertise and offer all jobs, including senior roles, as having flexible working options, such as part-time work, remote working, job sharing or compressed hours
-            </li>
-            <li>
-                Encourage senior leaders to role model working flexibly and to champion flexible working
-            </li>
-            <li>
-                Encourage and enable men to work flexibly, so that it is not seen as only a benefit for women
-            </li>
-            <li>
-                Talk to fathers about changing their working patterns when they have children, not just mothers
-            </li>
-        </ul>
-
-        <h3 class="heading-small">
-            Encourage the uptake of Shared Parental Leave
-        </h3>
-        <p>
-            The gender pay gap widens dramatically after women have children but this could be reduced if men and women were able to share childcare more equally. Shared Parental Leave and Pay enables working parents to share up to 50 weeks of leave and up to 37 weeks of pay in their child’s first year.
-        </p>
-        <p>
-            Encourage take up of Shared Parental Leave (see our <a href="https://sharedparentalleave.campaign.gov.uk/" rel="noopener"><span class="visually-hidden">Shared Parental Leave </span>guidance</a>). For example:
-        </p>
-        <ul class="list list-bullet">
-            <li>
-                Inform future fathers that it is their legal right to request Shared Parental Leave
-            </li>
-            <li>
-                Provide future parents guidance and personal support to understand the scheme
-            </li>
-            <li>
-                Share and promote examples of senior leaders who have taken Shared Parental Leave in your organization
-            </li>
-            <li>
-                Offer enhanced Shared Parental Pay at the same level as enhanced maternity pay
-            </li>
-        </ul>
-    </expand>
-</div>
-
-<Accordion>
-    <Section id="evidence-categories" title="Evidence categories">
-        <div class="action-step">
-            <i class="icon icon-effective"></i>
-            <h3 class="heading-small">
-                Effective actions
-            </h3>
-            <p>
-                There is strong evidence that shows these actions are effective, and that they are worth implementing. We recommend prioritising these actions above the actions in the categories below.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-promising"></i>
-            <h3 class="heading-small">
-                Promising actions
-            </h3>
-            <p>
-                These actions are promising but they still need further research to improve the evidence of their effectiveness and how best to implement them. The government has policies in place supporting some of these actions to improve women's participation in the workforce. The government is evaluating their effectiveness and we recommend that employers evaluate their actions too.
-            </p>
-        </div>
-        <div class="action-step">
-            <i class="icon icon-mixed"></i>
-            <h3 class="heading-small">
-                Actions with mixed results
-            </h3>
-            <p>
-                These actions have been shown sometimes to have a positive impact and other times a negative impact. This might be due to how they are implemented or other factors that we do not fully understand yet. As a result, investing in these actions is a more risky use of resources. We cannot, therefore, make a general recommendation that these are good ways to reduce gender inequality.
-            </p>
-        </div>
-    </Section>
-    <Section id="References" title="References">
-        <h3 class="heading-small">
-            Promising action: Improve workplace flexibility for men and women
-        </h3>
-        <p>
-            <a class="nowrap" href="https://timewise.co.uk/wp-content/uploads/2018/07/Timewise_Flexible_Jobs_-Index_2018.pdf" rel="noopener">The Timewise Flexible Jobs Index 2018<span class="visually-hidden"> (pdf)</span></a>
-        </p>
-
-        <h3 class="heading-small">
-            Promising action: Encourage the uptake of Shared Parental Leave
-        </h3>
-        <p>
-            <a class="nowrap" href="https://sharedparentalleave.campaign.gov.uk/" rel="noopener">Shared Parental Leave</a>
-        </p>
-
-    </Section>
-</Accordion>
-
-<Pagination>
-    <Back action="Talent" label="Talent management, learning & development"/>
-    <Next action="Reading" label="Further reading"/>
-</Pagination>
-<div style="margin-bottom: 25px; font-size: 16px!important">
-    <a rel="download track" class="print-link" data-track-category="relatedLinkClicked" data-track-action="1.1 Related content" data-track-label="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf" data-track-options="{&quot;dimension28&quot;:&quot;1&quot;,&quot;dimension29&quot;:&quot;Gender pay gap explained&quot;}" href="/public/assets/pdf/GEO_Actions_to_Close_the_Gender_Pay_Gap.pdf">Print entire guide<span class="visually-hidden"> (pdf)</span></a>
-</div>
