--- GenderPayGap.WebUI/Views/ActionHubOld/Index.cshtml
@@ -1,146 +1,0 @@
-﻿@using GovUkDesignSystem.GovUkDesignSystemComponents
-@using GovUkDesignSystem
-@using GenderPayGap.Core
-
-@{
-    ViewBag.Title = "Actions to close the gender pay gap - Gender pay gap service";
-    Layout = "~/Views/GovUkFrontend/GovUkFrontendLayout.cshtml";
-}
-
-@section BeforeMain {
-    @{
-        var crumbs = new List<CrumbViewModel>
-        {
-            new CrumbViewModel
-            {
-                Text = "Home",
-                Href = Url.Action("Index", "Viewing")
-            },
-            new CrumbViewModel
-            {
-                Text = "Actions to close the gender pay gap"
-            }
-        };
-    }
-
-    @(Html.GovUkBreadcrumbs(new BreadcrumbsViewModel
-    {
-        Crumbs = crumbs
-    }))
-}
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <span class="govuk-caption-l">Guidance</span>
-        <h1 class="govuk-heading-xl">Actions to close the gender pay gap</h1>
-    </div>
-</div>
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <p class="govuk-body-l">
-            Actions that are likely to improve recruitment and progression of women and reduce the gender pay gap.
-        </p>
-        <hr class="govuk-section-break govuk-section-break--xl govuk-section-break--visible govuk-!-margin-bottom-2">
-        <div class="govuk-!-margin-bottom-6">
-            <p class="govuk-body-s govuk-!-margin-bottom-0">
-                Published 1 December 2020
-            </p>
-
-            <p class="govuk-body-s govuk-!-margin-bottom-0">
-                From: <a class="govuk-link" href="https://www.gov.uk/government/organisations/government-equalities-office" rel="noopener">
-                    Government Equalities Office
-                </a>
-            </p>
-        </div>
-    </div>
-</div>
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <h4 class="govuk-caption-m">Contents</h4>
-
-        <ul class="govuk-list guidance-list-with-dashes">
-            <li>
-                <a class="govuk-link govuk-body-s" href="@Url.Action("Overview", "ActionHub")">Overview</a>
-            </li>
-            <li>
-                <a class="govuk-link govuk-body-s" href="@Url.Action("Effective", "ActionHub")">Effective actions</a>
-            </li>
-            <li>
-                <a class="govuk-link govuk-body-s" href="@Url.Action("Promising", "ActionHub")">Promising actions</a>
-            </li>
-            <li>
-                <a class="govuk-link govuk-body-s" href="@Url.Action("MixedResult", "ActionHub")">Actions with mixed results</a>
-            </li>
-        </ul>
-    </div>
-</div>
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <div class="govuk-!-margin-top-4">
-            @RenderBody()
-        </div>
-    </div>
-    <div class="govuk-grid-column-one-third">
-        <div class="guidance-related-content">
-            <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible related-content-section-break">
-
-            <h3 class="govuk-heading-s">Related content</h3>
-
-            <ul class="govuk-list related-content-list">
-                <li>
-                    <a class="govuk-link govuk-body-s" href="/public/assets/pdf/action-plan-guidance.pdf" rel="download">
-                        Four steps to developing a gender pay gap action plan
-                        <span class="govuk-visually-hidden"> (pdf)</span>
-                    </a>
-                </li>
-                <li>
-                    <a class="govuk-link govuk-body-s" href="@Url.Action("EightWaysOverview", "EightWaysGuidance")">
-                        Eight ways to understand your organisation's gender pay gap
-                    </a>
-                </li>
-                <li>
-                    <a class="govuk-link govuk-body-s" href="@Url.Action("SearchResults", "Viewing")">
-                        Find and compare employer’s gender pay gaps
-                    </a>
-                </li>
-                @if (FeatureFlagHelper.IsFeatureEnabled(FeatureFlag.ReportingStepByStep))
-                {
-                    <li>
-                        <a class="govuk-link govuk-body-s" href="@Url.Action("StepByStepStandalone", "ReportingStepByStep")">
-                            Prepare and report your gender pay gap: step-by-step guidance for employers
-                        </a>
-                    </li>
-                }
-            </ul>
-            <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
-            <div>
-                <h4 class="govuk-caption-m">Collection</h4>
-                <a class="govuk-link govuk-body-s" href="@Global.GpgGuidanceLink">
-                    Gender pay gap reporting
-                </a>
-            </div>
-        </div>
-        <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
-        <div class="guidance-related-content">
-
-            <h3 class="govuk-heading-s">Further reading</h3>
-
-            <ul class="govuk-list related-content-list">
-                <li class="govuk-body-s">
-                    What Works: Gender equality by design<br />
-                    by Iris Bohnet<br />
-                    Harvard University Press, 2016
-                </li>
-                <li class="govuk-body-s">
-                    Why diversity programs fail<br />
-                    by F. Dobbin and A. Kalev<br />
-                    Harvard Business Review, 2016
-                </li>
-            </ul>
-        </div>
-
-    </div>
-</div>
