+++ GenderPayGap.WebUI/appsettings.json
@@ -11,7 +11,6 @@
     "Enrich": [ "FromLogContext" ]
   },
 
-  "ActionHubSwitchOverDate": "2100-01-01",
   "DaysToKeepBackupFiles": 35,
   "FeatureFlagReportingStepByStep": "false",
   "FeatureFlagPrivateManualRegistration": "false",
